+++ src/TheLongestYear/Integration/IGenericModConfigMenuApi.cs
@@ -20,5 +20,11 @@ namespace TheLongestYear.Integration
             Func<string, string> formatAllowedValue = null, string fieldId = null);
         void AddSectionTitle(IManifest mod, Func<string> text, Func<string> tooltip = null);
         void AddParagraph(IManifest mod, Func<string> text);
+        void AddNumberOption(IManifest mod, Func<int> getValue, Action<int> setValue, Func<string> name,
+            Func<string> tooltip = null, int? min = null, int? max = null, int? interval = null,
+            Func<int, string> formatValue = null, string fieldId = null);
+        void AddNumberOption(IManifest mod, Func<float> getValue, Action<float> setValue, Func<string> name,
+            Func<string> tooltip = null, float? min = null, float? max = null, float? interval = null,
+            Func<float, string> formatValue = null, string fieldId = null);
     }
 }